Create Discrete-Interval-Trendlines-with-Persistence
This program divides the trading day into discrete intervals and draws an upper trend line and lower trend line. The interval length can be adjusted as needed. In this program, when a new time interval is detected, the script saves that intervals high and low; once at least two intervals have been recorded, the script draws two trendiness using the previous two intervals' values. This script with others in this repository are constantly being updated. My personal and private repository is called the "money printer", ask me about it.