Create Primary-Intermediate-Short_Term-Analysis
Still a work in progress. The goal of this script is to overlay three rectangles encompassing the current month, week, and day high to low. For options trading this is my prescribed primary, intermediate, and short term analysis ranges. This is a bigger part of my personal private "money printer" repository. Ask me about it.