Create Dynamic-Weekly-Rectangles
This script is complete given its purpose to build rectangles encapsulating the ongoing weeks high and low value adapting to changes in these values. This provides me with the range of key levels that the given market conditions are performing in for the current and ongoing week. This is a part of my personal private "money printer" repository. ask me about it.