Dynamic-Daily-Rectangles
This script creates an ongoing rectangle encompassing the days high and low key values to determine the current trading range given the days market conditions. This is a short-term analysis technique. I look for signals that will increase the height of this rectangle to expand current highs or lows. This is a ver dynamic rectangle that provides a lot of information. This can include whether the market is going to be testing new current highs or lows, or whether traders will stay in the rectangle, waiting for bears or bulls to establish a dominant dynamic. This script is a part of my person private "money printer" repository. Ask me about it.