Dynamic-Monthly-Rectangles
This script constructs rectangles that encompass ongoing highs and lows across a month interval. This technical analysis serves as a North Star for weekly trading patterns. After about 2 weeks, this analysis can provide more than enough data to provide me with a range of true key levels in the market. If the highs and lows remain the same for an extended period of time in this analysis, we can expect there will be greater emphasis on volatility when the highs and lows from this analysis get tested. This script is part of my personal private "money printer" repository. Ask me about it.